+++ src/types.ts
@@ -2,6 +2,11 @@ export type IndexedObservation = Observation & {
     index: number;
 };
 
+export type Kokokor = {
+    content: string;
+    footnotes?: string;
+};
+
 export type Observation = {
     readonly bbox: BoundingBox;
     readonly text: string;
