+++ src/utils/indexing.ts
@@ -10,7 +10,7 @@ import type { IndexedObservation, Observation } from '@/types';
  * @param dpi            image DPI (defaults to 72)
  * @param pixelTolerance extra vertical slack in “pixels at 72dpi”
  */
-export const markObservationsWithIndex = (observations: Observation[], dpi: number, pixelTolerance = 5) => {
+export const indexObservationsAsLines = (observations: Observation[], dpi: number, pixelTolerance = 5) => {
     // how many device‐pixels of slack at this DPI?
     const effectiveYTolerance = pixelTolerance * (dpi / 72);
 
@@ -41,4 +41,46 @@ export const markObservationsWithIndex = (observations: Observation[], dpi: numb
 
     // finally, ensure grouped by line then y
     return marked.sort((a, b) => (a.index !== b.index ? a.index - b.index : a.bbox.y - b.bbox.y));
+};
+
+export const indexObservationsAsParagraphs = (
+    observations: Observation[],
+    verticalJumpFactor = 2,
+    widthTolerance = 0.85,
+) => {
+    // ensure top→bottom order
+    const byY = observations.toSorted((a, b) => a.bbox.y - b.bbox.y);
+
+    // precompute width threshold
+    const maxWidth = Math.max(...byY.map((o) => o.bbox.width));
+    const thresholdWidth = maxWidth * widthTolerance;
+
+    const out: IndexedObservation[] = [];
+    let index = 0;
+
+    for (let i = 0; i < byY.length; i++) {
+        const o = byY[i];
+
+        // check for big vertical jump
+        if (i > 1) {
+            const prev = byY[i - 1];
+            const prevPrev = byY[i - 2];
+            const gap = o.bbox.y - prev.bbox.y;
+            const prevGap = prev.bbox.y - prevPrev.bbox.y;
+            if (gap > prevGap * verticalJumpFactor) {
+                index++;
+            }
+        }
+
+        // tag this line
+        out.push({ ...o, index });
+
+        // if it’s a “short” line, bump to new paragraph afterwards
+        if (o.bbox.width < thresholdWidth) {
+            index++;
+        }
+    }
+
+    // final sort by paragraph then y (just to keep stable ordering)
+    return out.toSorted((a, b) => (a.index !== b.index ? a.index - b.index : a.bbox.y - b.bbox.y));
 };
