+++ src/utils/normalization.ts
@@ -4,7 +4,7 @@ export const mapOcrResultToRTLObservations = (observations: Observation[], image
     return observations.map((o) => ({ ...o, bbox: { ...o.bbox, x: imageWidth - o.bbox.x - o.bbox.width } }));
 };
 
-export function normalizeObservationsX(observations: Observation[], dpi: number, standardDPI = 300) {
+export const normalizeObservationsX = (observations: Observation[], dpi: number, standardDPI = 300) => {
     const thresholdPx = (standardDPI / dpi) * 5;
     const minX = Math.min(...observations.map((o) => o.bbox.x));
 
@@ -17,4 +17,6 @@ export function normalizeObservationsX(observations: Observation[], dpi: number,
 
         return o;
     });
-}
+};
+
+export const applyFooter = (observations: Observation[], footer: Observation) => {};
